% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reconnect_functions.R
\name{alignst}
\alias{alignst}
\title{align raster "orig" with raster "target"}
\usage{
alignst(orig = NULL, target = NULL, method = "near")
}
\arguments{
\item{orig}{a raster layer that should be modified}

\item{target}{a raster layer to which the orig should be matched}

\item{method}{a character value ("near","bilinear","cubic","cubicspline") for the method used for estimating new cell values in the terra::project function}
}
\value{
a raster layer that matches to the "target" layer in coordinate reference system, extent and resolution.
}
\description{
change the coordinate reference system, extent and resolution of "orig" to match the one of "target". The function relies on the 'sf' and 'terra'packages
}
