% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONN_functions.R
\name{centr_igraph}
\alias{centr_igraph}
\title{Igraph centrality measures}
\usage{
centr_igraph(
  x = NULL,
  id = NULL,
  resi = NULL,
  pa = NULL,
  mdist = NULL,
  weighted = TRUE,
  cutoffpr = 0.01,
  MST = FALSE,
  normalized = TRUE,
  mode = c("undirected"),
  areafun = sf::st_area,
  dist_type = "edge",
  cost_type = "euclidean",
  trans_fun = function(x) 1/mean(x),
  neighb = 8,
  alpha = 500,
  dispfop = "negex",
  savememory = FALSE
)
}
\arguments{
\item{x}{habitat raster or shape file, should be binary if raster: 1= habitat and 0= non-habitat}

\item{id}{character string, sfp geometry id column describing habitat patch id's}

\item{resi}{resistance raster file, should have the same crs and extent as sf}

\item{pa}{numeric vector of areas of habitat patches 1-n}

\item{mdist}{square matrix of pairwise distances between habitat patches 1-n}

\item{areafun}{function defining how patch areas are calculated in case the input is a raster or shapefile (default: sf::st_area)}

\item{dist_type}{character, distance type: either "edge" or "centroid" (default: "edge").}

\item{cost_type}{character, algorithm applied to calculate distances: either "euclidean" (default), "least-cost" or "commute-time"}

\item{trans_fun}{function expression, argument for the gdistance::transition() function, applied when moving between resi raster cells. Default for conductance: function(x) 1/mean(x)}

\item{neighb}{numeric, directions argument for the gdistance::transition() function: directions in which cells are connected (4, 8, 16, ...)}

\item{alpha}{A species-specific dispersal distance (e.g. average gap-crossing distance, in the units of the mdist values, default = 500)}

\item{dispfop}{An option to choose the dispersal survival function (dispfun parameter), currently 3 options implemented: "negex", "linear" and "log-sech"}
}
\value{
A list of \href{https://igraph.org/}{igraph} R-package derived graph centrality measures: betweenness centrality (BC), closeness centrality (CC) and node degree (ND)
}
\description{
calculates betweenness centrality (BC), closeness centrality (CC) and Node Degree (ND). This function works either with a habitat raster or shapefile (x), can include a resistance raster (resi) or a vector with patch area's (pa) and a distance matrix (mdist) see \href{https://www.sciencedirect.com/science/article/pii/S0169204607000709}{Saura & Pascual-Hortal(2007)}
}
