% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONN_functions.R
\name{costdist_pm}
\alias{costdist_pm}
\title{Pairwise cost distance matrix of habitat patches in a resistance landscape}
\usage{
costdist_pm(
  x = NULL,
  resi = NULL,
  id = NULL,
  areafun = sf::st_area,
  trans_fun = function(x) 1/mean(x),
  neighb = 8,
  dist_type = "edge",
  cost_type = "least-cost"
)
}
\arguments{
\item{x}{A raster object where habitat patches are set to 1 and matrix area is set to NA OR a shapefile object containing habitat patches}

\item{resi}{resistance raster file, should have the same crs and extent as sf}

\item{id}{character string, sfp geometry id column describing habitat patch id's}

\item{areafun}{function defining how patch areas are calculated in case the input is a raster or shapefile (default: sf::st_area)}

\item{trans_fun}{function expression, argument for the gdistance::transition() function, applied when moving between resi raster cells. Default for conductance: function(x) 1/mean(x)}

\item{neighb}{numeric, directions argument for the gdistance::transition() function: directions in which cells are connected (4, 8, 16, ...)}

\item{dist_type}{character, distance type: either "edge" or "centroid" (default: "edge").}

\item{cost_type}{character, gdistance algorithm applied to calculate distances: either "least-cost" or "commute-time" (default: "least-cost", the randomised shortest paths, "gdistance::rSPDistance()" is not implemented)}
}
\value{
A list with an sf object of the habitat patches (sfpa) and a corresponding square matrix (cosdi) where pairwise cost distances among habitat patches are described according to the above defined criteria
}
\description{
this is a wrapper function combining the nice suite of \href{https://cran.r-project.org/web/packages/gdistance/index.html}{gdistance} functions
}
