% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONN_functions.R
\name{dI_fun}
\alias{dI_fun}
\title{Patch Importance}
\usage{
dI_fun(
  fun = NULL,
  expr = NULL,
  x = NULL,
  pid = NULL,
  resnr = 1,
  aslist = FALSE
)
}
\arguments{
\item{fun}{a character naming the connectivity metric function name; e.g. fun = "PCnum"}

\item{expr}{a character string describing the fun statement. Importantly, the user needs to specify "k", i.e. how list objects should  be subsetted; e.g. expression = "pa=x[\link{1}][-k],mdist=x[\link{2}][-k,-k] ,alpha=500,dispfop='negex',savememory=FALSE"}

\item{x}{a list containing all objects needed by fun, e.g. x  = list(pa=pa,mdist=mdist), whereby pa and mdist are objects in the R-environment}

\item{pid}{character, give an id name, e.g. describing habitat patch id's, the length of pid should match the length of the first element of x (i.e. length(x[\link{1}]))}

\item{resnr}{numeric, in case the fun has several result outputs, specify the number of the result output one is interested}

\item{aslist}{logical: in case the results should  be returned as list and not as matrix (default = FALSE)}
}
\value{
A matrix with patch importance values in rows for each connectivity indicator in columns
}
\description{
calculates patch importance for any connectivity metric provided, according to the formulas in e.g. \href{https://www.sciencedirect.com/science/article/pii/S0169204607000709}{Saura & Pascual-Hortal(2007)} and \href{https://onlinelibrary.wiley.com/doi/full/10.1111/j.1600-0587.2009.05760.x}{Saura and Rubio (2010)}.
}
