% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONN_functions.R
\name{ECA_pc}
\alias{ECA_pc}
\title{Equivalent Connected Area index (ECA) based on the probability of connectivity (PCnum)}
\usage{
ECA_pc(
  pa = NULL,
  mdist = NULL,
  alpha = 500,
  dispfop = "negex",
  savememory = FALSE,
  pcnum = NULL
)
}
\arguments{
\item{pa}{A numeric vector of areas of habitat patches 1-n}

\item{mdist}{A square matrix of pairwise distances between habitat patches 1-n}

\item{alpha}{A species-specific dispersal distance (e.g. average gap-crossing distance, in the units of the mdist values, default = 500)}

\item{dispfop}{An option to choose the dispersal survival function (dispfun parameter), currently 3 options implemented: "negex", "linear" and "log-sech"}
}
\value{
A numeric value of the Equivalent Connected Area index (ECA) based on the probability of connectivity (PCnum)
}
\description{
The ECA, calculated as the  square root of the numerator of the probability of connectivity index (PCnum), can be interpreted as the size of a single, maximally connected habitat patch that would provide the same probability of connectivity value than the provided habitat configuration (cf. \href{https://doi.org/10.1016/j.ecolind.2010.06.011}{Saura, Estreguil, Mouton & Rodríguez-Freire 2011}).
}
