% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONN_functions.R
\name{eucdist_hab}
\alias{eucdist_hab}
\title{Euclidean distance from within habitat to habitat edge}
\usage{
eucdist_hab(x = NULL, exclude = NULL)
}
\arguments{
\item{x}{a raster object where habitat patches are set to 1 and matrix area is set to NA}

\item{exclude}{numeric value of the cells that should not be considered for computing distances, used by the terra::distance function.}
}
\value{
A list with an sf object of the habitat patches (sfpa) and a corresponding square matrix (cosdi) where pairwise cost distances among habitat patches are described according to the above defined criteria
}
\description{
this is a much simplified wrapper function for the terra::distance function
}
