% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONN_functions.R
\name{eucdist_pm}
\alias{eucdist_pm}
\title{Pairwise euclidean distance between habitat patches}
\usage{
eucdist_pm(x = NULL, id = NULL, areafun = sf::st_area, dist_type = "edge")
}
\arguments{
\item{x}{a raster or shapefile describing habitat patches}

\item{id}{character string, sfp geometry id column describing habitat patch id's}

\item{areafun}{function defining how patch areas are calculated in case the input is a raster or shapefile (default: sf::st_area)}

\item{dist_type}{character, distance type: either "edge" or "centroid" (default: "edge").}
}
\value{
A list with a vector of habitat patch areas (pa) and a distance matrix (mdist)
}
\description{
convert a raster or shapefile into patch area and mdist matrix objects
}
