% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reconnect_functions.R
\name{fillNAr}
\alias{fillNAr}
\title{fill NA values in raster as simple mean of surrounding cells}
\usage{
fillNAr(x = NULL, roi = NULL, nmax = 100, ws = 5)
}
\arguments{
\item{x}{raster layer with NAs}

\item{roi}{shapefile of region of interest (optional)}

\item{nmax}{maximum nr of iterations (in case window size is small and NA gaps are large)}

\item{ws}{numeric value, must be an odd number, describing maximum window size applied in "focal" function of the raster package}
}
\value{
a raster layer with filled NA gaps, derived from linear interpolation for NA cells in windows with iteratively decreasing sizes, starting with ws
}
\description{
simple linear interpolation of raster datagaps using the raster::focal() function
}
