% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONN_functions.R
\name{invCR_l}
\alias{invCR_l}
\title{Inverse of cumulative resistance in landscape}
\usage{
invCR_l(
  x = NULL,
  resi = NULL,
  scale = 1,
  maxiter = 100,
  cost_type = "euclidean",
  dist_type = "edge",
  exclude = NULL
)
}
\arguments{
\item{x}{habitat raster or shape file, should be binary if raster: 1= habitat and 0= non-habitat}

\item{resi}{resistance raster file, should have the same crs and extent as sf}

\item{scale}{numeric value: scale factor in the terra::costDist function - the cost distance is divided by this number}

\item{maxiter}{numeric value: maximum number of iterations used by the terra::costDist function}

\item{cost_type}{character, algorithm applied to calculate distances: either "euclidean" (default) or "least-cost"}

\item{dist_type}{character, distance type: either "edge" or "centroid" (default: "edge").}

\item{exclude}{numeric value of the cells that should not be considered for computing distances, used by the terra::distance function.}
}
\value{
A numeric value of the inverse of cumulative resistance in the landscape (invCR_l)
}
\description{
invCR_l is the inverse of the average cumulative resistance distance to habitat patches in a landscape
}
