% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONN_functions.R
\name{invCR_p}
\alias{invCR_p}
\title{Inverse of cumulative resistance in patch network}
\usage{
invCR_p(
  x = NULL,
  id = NULL,
  resi = NULL,
  pa = NULL,
  mdist = NULL,
  pnr = 1,
  areafun = sf::st_area,
  dist_type = "edge",
  cost_type = "least-cost",
  trans_fun = function(x) 1/mean(x),
  neighb = 8,
  savememory = FALSE
)
}
\arguments{
\item{x}{habitat raster or shape file, should be binary if raster: 1= habitat and 0= non-habitat}

\item{id}{character string, sfp geometry id column describing habitat patch id's}

\item{resi}{resistance raster file, should have the same crs and extent as sf}

\item{pa}{numeric vector of areas of habitat patches 1-n}

\item{mdist}{square matrix of pairwise distances between habitat patches 1-n}

\item{pnr}{numeric, number of desired samples taken from all patches to calculate average cumulative resistance (default=100). Indicate 1 if all patches should be selected.}

\item{areafun}{function defining how patch areas are calculated in case the input is a raster or shapefile (default: sf::st_area)}

\item{dist_type}{character, distance type: either "edge" or "centroid" (default: "edge").}

\item{cost_type}{character, algorithm applied to calculate distances: either "euclidean" (default), "least-cost" or "commute-time"}

\item{trans_fun}{function expression, argument for the gdistance::transition() function, applied when moving between resi raster cells. Default for conductance: function(x) 1/mean(x)}

\item{neighb}{numeric, directions argument for the gdistance::transition() function: directions in which cells are connected (4, 8, 16, ...)}
}
\value{
A numeric value of the inverse of cumulative resistance in patch network (invCR_p)
}
\description{
invCR is the inverse of the average cumulative resistance of the shortest path between p randomly selected pairs of habitat patches k (pairs are drawn p times; cf. \href{https://doi.org/10.1111/cobi.12943}{Albert et al. (2017)} )
}
