% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONN_functions.R
\name{LandStats}
\alias{LandStats}
\title{Make a selection of basic landscape statistical metrics}
\usage{
LandStats(
  x = NULL,
  id = NULL,
  resi = NULL,
  pa = NULL,
  mdist = NULL,
  areafun = sf::st_area,
  dist_type = "edge",
  cost_type = "euclidean",
  trans_fun = function(x) 1/mean(x),
  neighb = 8,
  savememory = FALSE
)
}
\arguments{
\item{x}{habitat raster or shape file, should be binary if raster: 1= habitat and 0= non-habitat}

\item{id}{character string, sfp geometry id column describing habitat patch id's}

\item{resi}{resistance raster file, should have the same crs and extent as sf}

\item{pa}{numeric vector of areas of habitat patches 1-n}

\item{mdist}{square matrix of pairwise distances between habitat patches 1-n}

\item{areafun}{function defining how patch areas are calculated in case the input is a raster or shapefile (default: sf::st_area)}

\item{dist_type}{character, distance type: either "edge" or "centroid" (default: "edge").}

\item{cost_type}{character, algorithm applied to calculate distances: either "euclidean" (default), "least-cost" or "commute-time"}

\item{trans_fun}{function expression, argument for the gdistance::transition() function, applied when moving between resi raster cells. Default for conductance: function(x) 1/mean(x)}

\item{neighb}{numeric, directions argument for the gdistance::transition() function: directions in which cells are connected (4, 8, 16, ...)}
}
\value{
A named vector of landscape statistical metrics: total habitat area (tpa), average and standard deviation of habitat area (mnpa and sdpa), total number of patches (tnp), total nearest neighbour distances (tnnd), average and standard deviation of nearest neighbour distances (mnnnd and sdnnd)
}
\description{
Calculate basic landscape statistics for euclidean and cost distance landscapes
}
