% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reconnect_functions.R
\name{minarm}
\alias{minarm}
\title{remove small patches}
\usage{
minarm(shp = NULL, minsize = 0)
}
\arguments{
\item{shp}{a shapefile with polygons}

\item{minsize}{numeric value of minimum patch area in units of shapefile}
}
\value{
a new shapefile where small patches have been removed
}
\description{
remove all habitat patches smaller than a minimum area
}
