% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reconnect_functions.R
\name{mkbuffer}
\alias{mkbuffer}
\title{make a square or circular buffer for a given coordinate}
\usage{
mkbuffer(centers, radius, form = c("square"))
}
\arguments{
\item{centers}{a sf object (including crs, a projection with unit = meters) with coordinates for center points of buffer locations}

\item{radius}{the radius of the circular buffer, or half the side length of the square buffer in meters}

\item{form}{the form of the buffer: currently only circular or square}
}
\value{
a sf object with corresponding buffers
}
\description{
make a buffer sf object around a given sf spatial points center point
}
