% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reconnect_functions.R
\name{mkcenters}
\alias{mkcenters}
\title{make centerpoints of potential windows of selection with given distance apart, within extent "ext" and where "incr" is subtracted}
\usage{
mkcenters(ext = NULL, incr = NA)
}
\arguments{
\item{ext}{a raster::extent object}

\item{incr}{numeric value defining distance among centerpoints}
}
\value{
a shapefile containing a regular grid of centerpoints within extent ext
}
\description{
get an sf object of by incr regularly spaced centerpoints in a given extent (ext)
}
