% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLMC_functions.R
\name{mkcentersi}
\alias{mkcentersi}
\title{make centerpoints of potential windows of selection with given distance apart, within extent where "incr" is subtracted}
\usage{
mkcentersi(ext = NULL, incr = NA)
}
\arguments{
\item{ext}{a raster::extent object}

\item{incr}{numeric value defining distance among centerpoints}
}
\value{
a shapefile containing a regular grid of centerpoints within ext
}
\description{
get an sf object of by incr regularly spaced centerpoints in a given extent (ext)
}
