% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reconnect_functions.R
\name{mkdispfun}
\alias{mkdispfun}
\title{define a dispersal probability function}
\usage{
mkdispfun(option = "negex")
}
\arguments{
\item{option}{four options are currently available: "negex" (negative exponential, default), "linear" (linear distance decay), "log-sech" (long-tail distance decay) and "quantile", (negative exponential with alpha being median \link{quantile=0.5} or maximum \link{quantile=0.1} dispersal distance, default quantile value = 0.5, should be >0 and <=1)}
}
\value{
a dispersal probability function describing the dispersal kernel for given distances (d) and dispersal capacities of species (alpha).
}
\description{
create a function that calculates the dispersal probability between two points separated by distance "d", for a hypothetical species with dispersal capacity "alpha"
}
