% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPC_functions.R
\name{MPCser}
\alias{MPCser}
\title{calculate a series of Metapopulation Capacity (MPC) - based metrics}
\usage{
MPCser(
  roi = NULL,
  MPCbasl = NULL,
  x = NULL,
  id = NULL,
  resi = NULL,
  pa = NULL,
  mdist = NULL,
  alpha = 500,
  dispfop = "negex",
  ex = 0.5,
  self = TRUE,
  evec = TRUE,
  symmetric = FALSE,
  lower = TRUE,
  areafun = sf::st_area,
  dist_type = "edge",
  cost_type = "euclidean",
  trans_fun = function(x) 1/mean(x),
  neighb = 8,
  savememory = FALSE
)
}
\arguments{
\item{roi}{shapefile of region of interest}

\item{MPCbasl}{numeric value of baseline MPCvalue}
}
\value{
a list with MPC based metrics: MPCraw = raw MPC value, MPCev = eigenvector for each patch ID, MPCdens=MPC density, MPCrmax=MPC relative to maximum potential.
}
\description{
it uses the same arguments as the MPC_fun with two additional ones: roi and a baseline MPCvalue
}
