% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONN_functions.R
\name{NNDf}
\alias{NNDf}
\title{extract nearest neighbour distance}
\usage{
NNDf(mdist)
}
\arguments{
\item{mdist}{a square distance matrix}
}
\value{
A numeric vector of nearest neigbour distance for each row in a distance matrix
}
\description{
get nearest neighbour distance (NND) for each row in a distance matrix
}
