% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reconnect_functions.R
\name{normalizef}
\alias{normalizef}
\title{normalizing function}
\usage{
normalizef(x, na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector of values to be normalized}
}
\value{
a numeric vector of normalized values
}
\description{
normalize values of a vector between 0-1
}
