% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONN_functions.R
\name{PCECA_fun}
\alias{PCECA_fun}
\title{Probability of Connectivity wrapper function}
\usage{
PCECA_fun(
  x = NULL,
  id = NULL,
  resi = NULL,
  pa = NULL,
  mdist = NULL,
  AL = NULL,
  areafun = sf::st_area,
  dist_type = "edge",
  cost_type = "euclidean",
  trans_fun = function(x) 1/mean(x),
  neighb = 8,
  alpha = 500,
  dispfop = "negex",
  savememory = FALSE
)
}
\arguments{
\item{x}{habitat raster or shape file, should be binary if raster: 1= habitat and 0= non-habitat}

\item{id}{character string, sfp geometry id column describing habitat patch id's}

\item{resi}{resistance raster file, should have the same crs and extent as sf}

\item{pa}{numeric vector of areas of habitat patches 1-n}

\item{mdist}{square matrix of pairwise distances between habitat patches 1-n}

\item{AL}{numeric value, necessary if resi=NULL; should describe landscape area in units^2 of pa}

\item{areafun}{function defining how patch areas are calculated in case the input is a raster or shapefile (default: sf::st_area)}

\item{dist_type}{character, distance type: either "edge" or "centroid" (default: "edge").}

\item{cost_type}{character, algorithm applied to calculate distances: either "euclidean" (default), "least-cost" or "commute-time"}

\item{trans_fun}{function expression, argument for the gdistance::transition() function, applied when moving between resi raster cells. Default for conductance: function(x) 1/mean(x)}

\item{neighb}{numeric, directions argument for the gdistance::transition() function: directions in which cells are connected (4, 8, 16, ...)}

\item{alpha}{A species-specific dispersal distance (e.g. average gap-crossing distance, in the units of the mdist values, default = 500)}

\item{dispfop}{An option to choose the dispersal survival function (dispfun parameter), currently 3 options implemented: "negex", "linear" and "log-sech"}
}
\value{
A list of indices derived from the probability of connectivity numerator "PCnum" in \href{https://onlinelibrary.wiley.com/doi/full/10.1111/j.1600-0587.2009.05760.x}{Saura & Rubio, 2010}. PC: Probability of Connectivity, ECA: Equivalent Connected Area index \href{https://doi.org/10.1016/j.ecolind.2010.06.011}{Saura, Estreguil, Mouton & Rodríguez-Freire 2011}, ECA_AL: ECA relative to maximum possible (if ECA_AL is multiplied with 100, it is the ProtConn index \href{https://doi.org/10.1016/j.ecolind.2016.12.047}{Saura et al. (2017)}), ECA_AP: ECA relative to habitat amount in landscape (fraction of habitat area that is effectively connected), AL: Landscape area, AP: total patch area
}
\description{
This function works either with a habitat raster or shapefile (x), can include a resistance raster (resi) or a vector with patch area's (pa) and a distance matrix (mdist) see \href{https://www.sciencedirect.com/science/article/pii/S0169204607000709}{Saura & Pascual-Hortal(2007)}
}
