% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONN_functions.R
\name{PCnum}
\alias{PCnum}
\title{Probability of Connectivity numerator (PCnum)}
\usage{
PCnum(
  pa = NULL,
  mdist = NULL,
  alpha = 500,
  dispfop = "negex",
  savememory = FALSE
)
}
\arguments{
\item{pa}{A numeric vector of areas of habitat patches 1-n}

\item{mdist}{A square matrix of pairwise distances between habitat patches 1-n}

\item{alpha}{A species-specific dispersal distance (e.g. average gap-crossing distance, in the units of the mdist values, default = 500)}

\item{dispfop}{An option to choose the dispersal survival function (dispfun parameter), currently 3 options implemented: "negex", "linear" and "log-sech".}
}
\value{
A numeric value of the probability of connectivity numerator value "PCnum" in Saura & Rubio (2010).
}
\description{
Calculates the numerator of the probability of connectivity index (PC) which is the probability that two points randomly placed in a landscape fall into habitat areas that are interconnected/reachable from each other (cf. \href{https://www.sciencedirect.com/science/article/pii/S0169204607000709}{Saura & Pascual-Hortal(2007)}, \href{https://onlinelibrary.wiley.com/doi/full/10.1111/j.1600-0587.2009.05760.x}{Saura and Rubio (2010)} ).
}
