% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONN_functions.R
\name{perimf}
\alias{perimf}
\title{extract patch perimeter}
\usage{
perimf(shp)
}
\arguments{
\item{shp}{a shapefile with polygons}
}
\value{
A numeric vector of perimeter of each polygon in shapefile
}
\description{
get perimeters of each polygon in a shapefile
}
