% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLMC_functions.R
\name{randclust_gen}
\alias{randclust_gen}
\title{Generate random cluster landscapes}
\usage{
randclust_gen(
  simname = "test",
  ncs = 500,
  hfrs = 0.17,
  cfs = 1,
  reps = 1,
  respath = NULL,
  ncores = 1
)
}
\arguments{
\item{simname}{character string, name of simulation run}

\item{ncs}{numeric value defining number of cells in x and y dimension of the square raster (default = 500, algorithm runs faster the smaller ncs)}

\item{hfrs}{numeric vector defining desired fraction of habitat in the landscape}

\item{cfs}{numeric vector defining the desired "clumping factor" realized by the proportion of elements randomly selected to form clusters: small values give more fine scale heterogeneity}

\item{reps}{numeric value indicating number of replicates per run}

\item{respath}{character string indicating file path where results should be saved}

\item{ncores}{number of cores to be used when parallel processing}
}
\value{
a raster file containing a landscape with habitat patches with above defined characteristics, saved to respath/simname
}
\description{
wrapper function applying the random cluster algorithm developped by developped by \href{https://link.springer.com/article/10.1023/A:1008107902848}{Saura and Martinez-Millan 2000} and implemented in the \href{https://github.com/ropensci/NLMR/}{NLMR package} by \href{https://besjournals.onlinelibrary.wiley.com/doi/10.1111/2041-210X.13076}{Sciaini et al. 2018}.
}
