% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reconnect_functions.R
\name{readini}
\alias{readini}
\title{read in the excel Reconnect inifile}
\usage{
readini(path = NULL, sheet = 1)
}
\arguments{
\item{path}{character specifying path to the ini file ("C:/user/mypath/myfile.xlsx")}

\item{sheet}{numeric value of the inifile excel sheet, default = 1}
}
\value{
a dataframe with the specified connectivity functions
}
\description{
read in the Reconnect inifile
}
