% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLMC_functions.R
\name{reallands_gen}
\alias{reallands_gen}
\title{Generate real-world landscapes with habitat distribution from an input raster file}
\usage{
reallands_gen(
  simname = "test",
  rast = NULL,
  ncs = 100,
  samplesize = 50,
  overlap = 0.5,
  respath = NULL,
  lcdf = NULL,
  habnames = NULL,
  lccol = "hab"
)
}
\arguments{
\item{simname}{character, name of simulation run}

\item{rast}{land cover raster file containing codes corresponding to the "Code" column in lcdf.}

\item{ncs}{numeric value defining number of cells in x and y dimension of the square raster to be extractedx}

\item{samplesize}{numeric value indicating the number of desired sample landscapes}

\item{overlap}{numeric value indicating the allowed overlap among sampled landscapes (1=no overlap,if <1 overlap will be bigger, if >1 distance among sampled landscapes will be larger)}

\item{respath}{character string indicating file path where results should be saved}

\item{lcdf}{data frame containing a "Code" column describing codes contained in rast; and a lccol with associated habitat names}

\item{habnames}{character string defining land cover names to be used to define habitat pixels in rast, habnames must be contained in the lccol of the lcdf dataframe.}

\item{lccol}{character string defining column name of lcdf where habitat names are contained}
}
\value{
a raster and/or a shapefile containing a landscape with habitat patches with above defined characteristics
}
\description{
wrapper function sampling landscapes of a certain size and for a given land cover value
}
