% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reconnect_functions.R
\name{Reconnect_core}
\alias{Reconnect_core}
\title{Function within Reconnect_wrap: apply connectivity functions in the "Reconnect_inifile.xlsx" for a given landscape}
\usage{
Reconnect_core(
  cid = NULL,
  pidname = NULL,
  hr = NULL,
  foc = NULL,
  hshp = NULL,
  minsizev = NULL,
  cutpatches = TRUE,
  cost_type = "euclidean",
  dist_type = "edge",
  trans_fun = function(x) 1/mean(x),
  neighb = 8,
  areafun = sf::st_area,
  cond = NULL,
  nplot = 10,
  cols = terrain.colors(100),
  maxnpatch = 28500,
  radius = NULL,
  resolution = NULL,
  incr = NULL,
  respath = NULL,
  habi = NULL
)
}
\arguments{
\item{cid}{character of the buffer-center shapefile id-column}

\item{pidname}{character describing the patch-id column name in the habitat patch shapefile (hshp)}

\item{hr}{raster layer or raster stack object, containing a habitat layer and a resistance layer.}

\item{foc}{shapefile, of focal region of interest}

\item{hshp}{shapefile of habitat patches}

\item{minsizev}{numeric vector of minimum habitat patch areas in m2}

\item{cutpatches}{logical, should habitat patches be cut at the boundary of the focal region of interest (foc)? Default = TRUE}

\item{cost_type}{character, algorithm applied to calculate distances: either "euclidean" (default), "least-cost" or "commute-time"}

\item{dist_type}{character, distance type: either "edge" or "centroid" (default: "edge").}

\item{trans_fun}{function expression, argument for the gdistance::transition() function, applied when moving between resi raster cells. Default for conductance: function(x) 1/mean(x)}

\item{neighb}{numeric, directions argument for the gdistance::transition() function: directions in which cells are connected (4, 8, 16, ...)}

\item{areafun}{function defining how patch areas are calculated in case the input is a raster or shapefile (default: sf::st_area)}

\item{cond}{dataframe, connectivity functions ini file (cf. 1-Data folder & readini function)}

\item{nplot}{numeric value defining the frequency with which plotting is done for the buffer id's (default: 10, i.e. every 10th id will be plotted)}

\item{cols}{color ramp palette to be used (default: terrain.colors(100))}

\item{maxnpatch}{numeric, maximum number of patches to be treated in a landscape (so that connectivity functions don't get overwhelmed)}

\item{radius}{numeric, radius of buffer window}

\item{resolution}{numeric, resolution of raster grid}

\item{incr}{numeric, incremental distance among buffer windows}

\item{respath}{character string indicating file path where results should be saved}
}
\value{
csv files saved to respath location, where results from the application of connectivity indicators specified in the cond file are saved (dfla = landscape-level, dfpa = patch-level, dfpi = pixel-level).
}
\description{
this function is applied in the Reconnect_wrap function, to calculate connectivity indices for a buffer window around a given sf spatial points center point
}
