% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reconnect_functions.R
\name{Reconnect_single}
\alias{Reconnect_single}
\title{Function independent of Reconnect_wrap: apply connectivity functions in the "Reconnect_inifile.xlsx" for a given landscape}
\usage{
Reconnect_single(
  cid = NULL,
  pidname = NULL,
  lcr = NULL,
  lres = NULL,
  foc = NULL,
  hshp = NULL,
  minsizev = 0,
  cutpatches = TRUE,
  cost_type = "euclidean",
  dist_type = "edge",
  trans_fun = function(x) 1/mean(x),
  neighb = 8,
  areafun = sf::st_area,
  cond = NULL,
  cols = terrain.colors(100),
  maxnpatch = 28500,
  resolution = NULL,
  respath = NULL
)
}
\arguments{
\item{cid}{character name or ID string}

\item{pidname}{character describing the patch-id column name in the habitat patch shapefile (hshp)}

\item{lcr}{raster habitat layer where habitat > 1 and non-habitat = 0 or NA.}

\item{lres}{raster layer of resistance surface}

\item{foc}{shapefile, of focal region of interest (optional)}

\item{hshp}{shapefile of habitat patches (optional)}

\item{minsizev}{numeric vector of minimum habitat patch areas in m2}

\item{cutpatches}{logical, should habitat patches be cut at the boundary of the focal region of interest (foc)? Default = TRUE}

\item{cost_type}{character, algorithm applied to calculate distances: either "euclidean" (default), "least-cost" or "commute-time"}

\item{dist_type}{character, distance type: either "edge" or "centroid" (default: "edge").}

\item{trans_fun}{function expression, argument for the gdistance::transition() function, applied when moving between resi raster cells. Default for conductance: function(x) 1/mean(x)}

\item{neighb}{numeric, directions argument for the gdistance::transition() function: directions in which cells are connected (4, 8, 16, ...)}

\item{areafun}{function defining how patch areas are calculated in case the input is a raster or shapefile (default: sf::st_area)}

\item{cond}{dataframe, connectivity functions ini file (cf. 1-Data folder & readini function)}

\item{cols}{color ramp palette to be used (default: terrain.colors(100))}

\item{maxnpatch}{numeric, maximum number of patches to be treated in a landscape (so that connectivity functions don't get overwhelmed)}

\item{resolution}{numeric, resolution of raster grid}

\item{respath}{character string indicating file path where results should be saved}
}
\value{
csv files saved to respath location, where results from the application of connectivity indicators specified in the cond file are saved (dfla = landscape-level, dfpa = patch-level, dfpi = pixel-level).
}
\description{
simple Reconnect core function: apply connectivity functions in the "Reconnect_inifile.xlsx" for a given landscape, without moving windows
}
