% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reconnect_functions.R
\name{Reconnect_summary}
\alias{Reconnect_summary}
\title{Summarize moving buffer results from the Reconnect_wrap function into manageable formats}
\usage{
Reconnect_summary(
  sourcepath = NULL,
  respath = NULL,
  habnames = "test",
  resolution = NULL,
  radius = NULL,
  incr = NA,
  mosafun = base::min,
  lcr0 = NULL,
  cutpatches = TRUE,
  cols = topo.colors,
  varset = c("NND"),
  valset = c("max")
)
}
\arguments{
\item{sourcepath}{character string indicating file path where Reconnect_wrap outputs are saved}

\item{respath}{character string indicating file path where results should be saved}

\item{habnames}{character string, indicating the habitat names of interest (same as used in Reconnect_wrap function)}

\item{resolution}{numeric, resolution of raster grid}

\item{radius}{numeric, radius of circular moving window (if moving window is square, radius = 0.5*window sidelength}

\item{incr, }{numeric, incremental distance between moving window centers}

\item{mosafun}{R-function used for generating a raster mosaic for the pixel-level connectivity functions}

\item{lcr0}{raster layer, original large raster layer that was used as input in the Reconnect_wrap function}

\item{cutpatches}{logical, should habitat patches be cut at the boundary of the focal region of interest (foc)? Default = TRUE}

\item{cols}{color pallette to be used when plotting (default = topo.colors)}

\item{varset}{character string of landscape-level connectivity variables that should be set to another value than 0 (default) in case of NA (e.g. variables where connectivity is highest at low values, such as nearest neighbour distance NND)}

\item{valset}{character or numeric string of values (current options if character: "mean","min","max","radius") to which landscape-level connectivity variables in varset should be set (order of valset must match varset).}
}
\value{
raster files saved to respath (landscape & pixel -level connectivity), shapefiles saved to respath (patch-level connectivity), and plots depicting connectivity at pixel-, patch- and landscape-level.
}
\description{
synthesize all moving buffer windows into 1 file for the grand region of interest (landscape & pixel level connectivity will be summarized as raster files, patch-level connectivity will be summarized as a shapefile)
}
