% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reconnect_functions.R
\name{Reconnect_wrap}
\alias{Reconnect_wrap}
\title{wrapper function for the Reconnect_core function}
\usage{
Reconnect_wrap(
  lcr = NULL,
  lres = NULL,
  roi = NULL,
  cond = NULL,
  respath = NULL,
  resolution = NULL,
  radius = NULL,
  incr = NA,
  habi = "test",
  bufform = "square",
  minsizev = c(0),
  cutpatches = TRUE,
  ncores = 1,
  oviewplot = TRUE,
  cost_type = "euclidean",
  dist_type = "edge",
  trans_fun = function(x) 1/mean(x),
  neighb = 8,
  areafun = sf::st_area,
  maxnpatch = 28500,
  cst = NULL,
  cen = NULL,
  hcols = rev(c("yellow", "orange", "red", "purple", "blue", "lightblue")),
  rcols = rev(c("purple", "lightblue", "yellow", "olivedrab", "darkgreen"))
)
}
\arguments{
\item{lcr}{raster layer object, containing a habitat patches (binary, habitat = 1, other = NA)}

\item{lres}{raster layer object, containing a resistance surface}

\item{roi}{shapefile, region of interest}

\item{cond}{dataframe, connectivity functions ini file (cf. 1-Data folder & readini function)}

\item{respath}{character string indicating file path where results should be saved}

\item{resolution}{numeric, resolution of raster grid}

\item{radius}{numeric, radius of circular moving window (if moving window is square, radius = 0.5*window sidelength}

\item{incr, }{numeric, incremental distance between moving window centers}

\item{habi}{character, habitat name}

\item{bufform}{character, form of buffer window used by the mkbuffer function (default: "square", alternative "circle")}

\item{minsizev}{numeric vector of minimum habitat patch areas in m2}

\item{cutpatches}{logical, should habitat patches be cut at the boundary of the focal region of interest (foc)? Default = TRUE}

\item{ncores}{numeric, number of cores that should be used in the parallel processing of moving buffers}

\item{oviewplot}{logical, should an overview plot be generated or not? (default: TRUE)}

\item{cost_type}{character, algorithm applied to calculate distances: either "euclidean" (default), "least-cost" or "commute-time"}

\item{dist_type}{character, distance type: either "edge" or "centroid" (default: "edge").}

\item{trans_fun}{function expression, argument for the gdistance::transition() function, applied when moving between resi raster cells. Default for conductance: function(x) 1/mean(x)}

\item{neighb}{numeric, directions argument for the gdistance::transition() function: directions in which cells are connected (4, 8, 16, ...)}

\item{areafun}{function defining how patch areas are calculated in case the input is a raster or shapefile (default: sf::st_area)}

\item{maxnpatch}{numeric, maximum number of patches to be treated in a landscape (so that connectivity functions don't get overwhelmed)}

\item{cst}{numeric, start of moving window indices that should be computed (if NULL as in default, cst will be set to 1)}

\item{cen}{numeric, end of moving window indices that should be computed (if NULL as in default, cen will be set to the maximum nr of moving windows existent)}

\item{hcols}{character string of colors used in habitat layer plotting}

\item{rcols}{character string of colors used in resistance layer plotting}
}
\value{
csv files saved to respath location, where results from the application of connectivity indicators specified in the cond file are saved (dfla = landscape-level, dfpa = patch-level, dfpi = pixel-level).
}
\description{
subsets your habitat raster file into bufferzones, prepares data such that connectivity indices can be calculated in a moving buffer using the Reconnect_core function and connectivity functions specified in the Reconnect_ini file.
}
