% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONN_functions.R
\name{rstosf}
\alias{rstosf}
\title{convert habitat raster to shapefile with habitat patches}
\usage{
rstosf(rs = NULL)
}
\arguments{
\item{rs}{a raster file describing habitat patches}
}
\value{
A shapefile with patch ids in column "pid"
}
\description{
convert a binary habitat distribution raster file into a shapefile
}
