% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLMC_functions.R
\name{simpcirc}
\alias{simpcirc}
\title{Simple circle habitat patch simulation}
\usage{
simpcirc(
  dimx = 500,
  dimy = 500,
  hfr = 0.17,
  npatch = 5,
  sdpa = 0,
  cf = 1,
  form = "circle",
  return = "raster",
  remedge = TRUE
)
}
\arguments{
\item{dimx}{number of raster cells in the x dimension}

\item{dimy}{number of raster cells in the y dimension}

\item{hfr}{desired fraction of habitat in within the whole landscape (note that the larger hfr, the longer it takes to randomly sample habitat patch locations, especially if the clumping factor cf < 1, i.e. if they should be spread out)}

\item{npatch}{desired number of habitat patches in the landscape}

\item{sdpa}{desired standard deviation of habitat patch areas - if sd = 0 (default), all habitat patches have the same size}

\item{cf}{desired clumping factor of habitat patch distribution - if cf = 1 (default), habitat patches are non-overlapping; if cf<1 patches are sampled to be further apart, if cf>1,patches are sampled to be closer and overlapping, likely leading to a reduction in total habitat area, lower desired npatch and different forms of habitat patches.}

\item{form}{desired form of the habitat patches, currently available: "circle" (default) and "square".}

\item{return}{return a raster file ("raster",default), a shapefile ("shapefile") or a raster and shapefile ("all")?}

\item{remedge}{logical -  should samples have a minimum distance (mindist) to the raster edge (TRUE) or not (FALSE, default)}
}
\value{
a raster and/or a shapefile containing a landscape with habitat patches with above defined characteristics
}
\description{
simulate landscapes with defined number of patches, fraction of habitat (note: because of the random generation process, landscapes can slightly deviate from desired settings in hfr, npatch, sdpa and cf)
}
