% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLMC_functions.R
\name{simpcirc_gen}
\alias{simpcirc_gen}
\title{Generate simple circle raster landscapes}
\usage{
simpcirc_gen(
  simname = "test",
  ncs = 500,
  hfrs = 0.17,
  npatches = c(1:10),
  sdpas = 0,
  cfs = 1,
  reps = 1,
  respath = NULL,
  remedge = TRUE
)
}
\arguments{
\item{simname}{character, name of simulation run}

\item{ncs}{numeric value defining number of cells in x and y dimension of the square raster (default = 500)}

\item{hfrs}{numeric vector defining desired fraction of habitat in within the whole landscape (note that the larger hfr, the longer it takes to randomly sample habitat patch locations, especially if the clumping factor cf < 1, i.e. if they should be spread out)}

\item{npatches}{numeric vector defining desired number of habitat patches in the simulated landscapes}

\item{sdpas}{numeric vector defining desired standard deviation of habitat patch areas - if sd = 0 (default), all habitat patches have the same size}

\item{cfs}{numeric vector defining desired clumping factor of habitat patch distribution - if cf = 1 (default), habitat patches are non-overlapping; if cf<1 patches are sampled to be further apart, if cf>1,patches are sampled to be closer and overlapping, likely leading to a reduction in total habitat area, lower desired npatch and different forms of habitat patches.}

\item{reps}{numeric value indicating number of replicates per run}

\item{respath}{character string indicating file path where results should be saved}

\item{remedge}{logical -  should samples have a minimum distance (mindist) to the raster edge (TRUE) or not (FALSE, default)}
}
\value{
a raster and/or a shapefile containing a landscape with habitat patches with above defined characteristics
}
\description{
wrapper function applying the simpcirc function: simulate landscapes with defined number of patches, fraction of habitat (note: because of the random generation process, landscapes can slightly deviate from desired settings in hfr, npatch, sdpa and cf)
}
