% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLMC_functions.R
\name{stratsamp}
\alias{stratsamp}
\title{stratified random sampling in strata of interest, with minimum distance argument}
\usage{
stratsamp(
  rast = NULL,
  strata = NULL,
  nsamp = 10,
  mindist = 100,
  remedge = FALSE,
  maxiter = 500
)
}
\arguments{
\item{rast}{a raster or raster stack containing the strata from which to sample}

\item{strata}{a character string describing the stratum raster layer name}

\item{nsamp}{numeric value indicating the desired number of replicates for each stratum type}

\item{mindist}{numeric value indicating the minimum distance among samples withing a stratum}

\item{remedge}{logical -  should samples have a minimum distance (mindist) to the raster edge (TRUE) or not (FALSE, default)}

\item{maxiter}{numeric value indicating the maximum number of iterations for searching the optimal sample set per stratum}
}
\value{
a list of 3 lists: idx = raster cell index values of samples per stratum, fracmind=fraction of samples undercutting mindist threshold, length=number of samples
}
\description{
sample an equal number of replicates for strata of interest, such as urban and non-urban land covers, with minimum distance criterion
}
