% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLMC_functions.R
\name{summplots}
\alias{summplots}
\title{Generate overview plots of simulation log data}
\usage{
summplots(
  simname = "test",
  respath = NULL,
  makepdf = FALSE,
  rdf = NULL,
  metrics = NULL,
  groups = NULL,
  cols = topo.colors,
  simpcirc = FALSE
)
}
\arguments{
\item{simname}{character, name of simulation run}

\item{respath}{character string indicating file path where results should be saved}

\item{makepdf}{logical, should a pdf of the plots be created? (default=FALSE)}

\item{rdf}{a results data frame, containing the column names "metric", "value", "simname" as well as names specified in the "groups" variable}

\item{metrics}{a character string defining metrics in the "metric" column that should be summarized}

\item{groups}{a character string indicating column names in "rdf" that serve as grouping variables when summarizing the data}

\item{cols}{color pallette to be used when plotting (default = topo.colors)}

\item{simpcirc}{logical - are data coming from the simpcirc simulation (simpcirc_gen function output)? Default = FALSE}
}
\value{
Plots depicting relationships among metrics, grouped by "groups" and saved as a pdf file in respath (if makepdf=TRUE)
}
\description{
plot relationships among metrics in simulated landscapes, grouped by groups
}
